Refactor mutable ast logic to its own extension class
