+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/ProjectExtension.kt
@@ -1,5 +1,6 @@
 package io.gitlab.arturbosch.detekt.api
 
+import io.gitlab.arturbosch.detekt.api.internal.DetektPomModel
 import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
 import org.jetbrains.kotlin.cli.common.config.addKotlinSourceRoots
 import org.jetbrains.kotlin.cli.common.messages.MessageRenderer
@@ -9,21 +10,12 @@ import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
 import org.jetbrains.kotlin.cli.jvm.config.addJavaSourceRoots
 import org.jetbrains.kotlin.cli.jvm.config.addJvmClasspathRoots
 import org.jetbrains.kotlin.com.intellij.mock.MockProject
-import org.jetbrains.kotlin.com.intellij.openapi.extensions.ExtensionPoint
-import org.jetbrains.kotlin.com.intellij.openapi.extensions.Extensions
 import org.jetbrains.kotlin.com.intellij.openapi.util.Disposer
-import org.jetbrains.kotlin.com.intellij.openapi.util.UserDataHolderBase
 import org.jetbrains.kotlin.com.intellij.pom.PomModel
-import org.jetbrains.kotlin.com.intellij.pom.PomModelAspect
-import org.jetbrains.kotlin.com.intellij.pom.PomTransaction
-import org.jetbrains.kotlin.com.intellij.pom.impl.PomTransactionBase
-import org.jetbrains.kotlin.com.intellij.pom.tree.TreeAspect
-import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.TreeCopyHandler
 import org.jetbrains.kotlin.config.CommonConfigurationKeys
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.config.JVMConfigurationKeys
 import org.jetbrains.kotlin.config.JvmTarget
-import sun.reflect.ReflectionFactory
 import java.io.File
 import java.nio.file.Path
 
@@ -39,9 +31,8 @@ fun createKotlinCoreEnvironment(configuration: CompilerConfiguration = CompilerC
         EnvironmentConfigFiles.JVM_CONFIG_FILES
     )
 
-    val mockProject = environment.project as? MockProject ?: error("MockProject class type expected")
-    makeMutable(mockProject)
-//    mockProject?.registerService(PomModel::class.java, DetektPomModel())
+    val project = environment.project as? MockProject ?: error("MockProject type expected")
+    project.registerService(PomModel::class.java, DetektPomModel(project))
 
     return environment
 }
@@ -71,33 +62,4 @@ fun createCompilerConfiguration(
         addKotlinSourceRoots(kotlinFiles)
         addJvmClasspathRoots(classpath.map { File(it) })
     }
-}
-
-fun makeMutable(project: MockProject) {
-    // Based on KtLint by Shyiko
-    val pomModel: PomModel = object : UserDataHolderBase(), PomModel {
-
-        override fun runTransaction(transaction: PomTransaction) {
-            (transaction as? PomTransactionBase)?.run()
-        }
-
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : PomModelAspect> getModelAspect(aspect: Class<T>): T? {
-            if (aspect == TreeAspect::class.java) {
-                // using approach described in https://git.io/vKQTo due to the magical bytecode of TreeAspect
-                // (check constructor signature and compare it to the source)
-                // (org.jetbrains.kotlin:kotlin-compiler-embeddable:1.0.3)
-                val constructor = ReflectionFactory.getReflectionFactory().newConstructorForSerialization(
-                    aspect, Any::class.java.getDeclaredConstructor())
-                return constructor.newInstance() as T
-            }
-            return null
-        }
-    }
-    val extensionPoint = "org.jetbrains.kotlin.com.intellij.treeCopyHandler"
-    val extensionClassName = TreeCopyHandler::class.java.name!!
-    arrayOf(Extensions.getArea(project), Extensions.getArea(null))
-        .filter { !it.hasExtensionPoint(extensionPoint) }
-        .forEach { it.registerExtensionPoint(extensionPoint, extensionClassName, ExtensionPoint.Kind.INTERFACE) }
-    project.registerService(PomModel::class.java, pomModel)
 }
